#2437 split phoebus probe integration into stand alone plugin
