+++ applications/phoebus/phoebus-plugins/org.csstudio.phoebus.integration.probe/icons/probe-16.png
