+++ applications/phoebus/phoebus-plugins/org.csstudio.phoebus.integration.probe/icons/probe-48.png
