+++ applications/phoebus/phoebus-plugins/org.csstudio.phoebus.integration.probe/icons/probe-add-16.png
