+++ applications/phoebus/phoebus-plugins/org.csstudio.phoebus.integration.probe/plugin.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.4"?>
+<plugin>
+   <extension
+         point="org.eclipse.ui.menus">
+      <menuContribution
+            allPopups="false"
+            locationURI="popup:org.csstudio.ui.menu.popup.processvariable">
+         <command
+               commandId="org.csstudio.phoebus.integration.probe.OpenPhoebusProbe"
+               icon="icons/probe-16.png"
+               label="Phoebus Probe"
+               style="push">
+         </command>
+      </menuContribution>
+      <menuContribution
+            allPopups="false"
+            locationURI="menu:utility">
+         <command
+               commandId="org.csstudio.phoebus.integration.probe.OpenPhoebusProbe"
+               icon="icons/probe-16.png"
+               label="Phoebus Probe"
+               style="push">
+         </command>
+      </menuContribution>
+   </extension>
+   <extension
+         point="org.eclipse.ui.commands">
+      <command
+            defaultHandler="org.csstudio.phoebus.integration.probe.OpenPhoebusProbe"
+            icon="icons/probe-16.png"
+            id="org.csstudio.phoebus.integration.probe.OpenPhoebusProbe"
+            name="ProbeLabel">
+      </command>
+   </extension>
+</plugin>
