+++ applications/phoebus/phoebus-plugins/org.csstudio.phoebus.integration.probe/src/org/csstudio/phoebus/integration/probe/Activator.java
@@ -1,0 +1,25 @@
+package org.csstudio.phoebus.integration.probe;
+
+import org.osgi.framework.BundleActivator;
+import org.osgi.framework.BundleContext;
+
+public class Activator implements BundleActivator {
+
+    private static BundleContext context;
+
+    // The plug-in ID
+    public static final String PLUGIN_ID = "org.csstudio.phoebus.integration";
+
+    static BundleContext getContext() {
+        return context;
+    }
+
+    public void start(BundleContext bundleContext) throws Exception {
+        Activator.context = bundleContext;
+    }
+
+    public void stop(BundleContext bundleContext) throws Exception {
+        Activator.context = null;
+    }
+
+}
