+++ applications/phoebus/phoebus-plugins/org.csstudio.phoebus.integration.probe/src/org/csstudio/phoebus/integration/probe/messages.properties
@@ -1,0 +1,11 @@
+#Eclipse modern messages class
+#Tue Mar 05 09:38:05 EST 2013
+MultipleInstancesFmt=You have {0} PV names selected.\nOpen an instance of Probe for each of them?
+MultipleInstancesTitle=Open multiple Probes?
+
+# Probe
+
+# Probe app name
+Probe=probe
+ProbeErrorOpenProbe=Cannot open probe
+ProbeLabel=Phoebus Probe
