+++ applications/phoebus/phoebus-plugins/pom.xml
@@ -11,5 +11,6 @@
     <module>org.csstudio.phoebus.integration</module>
     <module>org.csstudio.phoebus.integration.alarm</module>
     <module>org.csstudio.phoebus.integration.channel</module>
+    <module>org.csstudio.phoebus.integration.probe</module>
   </modules>
 </project>
